feat(Context): Added Context component that takes care of the positioning and arrow styles. See breaking changes.
BREAKING CHANGE: To allow a combination of ContextBox and ContextMeny styles to work together within the same context a new component called Context has been introduced. Simply wrap existing ContextBox and ContextMenu components with this one and transfer properties except padding prop.
