+++ src/components/context/Context.scss
@@ -1,0 +1,85 @@
+@import '../../design-patterns/colors/vars';
+@import '../../design-patterns/colors/colors-light';
+@import '../../design-patterns/spacing/vars';
+@import '../vars';
+
+$context-tip-size: $spacing-grid-size * 3;
+
+.ax-context {
+  position: relative;
+  filter: drop-shadow($drop-shadow);
+}
+
+.ax-context__content {
+  border-radius: $component-border-radius;
+  background-color: var(--color-context-background);
+  color: var(--color-context-text);
+  overflow: hidden;
+}
+
+.ax-context__arrow {
+  position: absolute;
+  width: 0;
+  height: 0;
+  border-width: $context-tip-size;
+  border-style: solid;
+  border-color: transparent;
+}
+
+.ax-context--top {
+  margin-bottom: $context-tip-size;
+}
+
+.ax-context--bottom {
+  margin-top: $context-tip-size;
+}
+
+.ax-context--right {
+  margin-left: $context-tip-size;
+}
+
+.ax-context--left {
+  margin-right: $context-tip-size;
+}
+
+.ax-context--top,
+.ax-context--bottom {
+  .ax-context__arrow {
+    left: calc(50% - #{$context-tip-size});
+  }
+}
+
+.ax-context--right,
+.ax-context--left {
+  .ax-context__arrow {
+    top: calc(50% - #{$context-tip-size});
+  }
+}
+
+.ax-context--top {
+  .ax-context__arrow {
+    bottom: -($context-tip-size * 2);
+    border-top-color: var(--color-context-background);
+  }
+}
+
+.ax-context--right {
+  .ax-context__arrow {
+    left: -($context-tip-size * 2);
+    border-right-color: var(--color-context-background);
+  }
+}
+
+.ax-context--bottom {
+  .ax-context__arrow {
+    top: -($context-tip-size * 2);
+    border-bottom-color: var(--color-context-background);
+  }
+}
+
+.ax-context--left {
+  .ax-context__arrow {
+    right: -($context-tip-size * 2);
+    border-left-color: var(--color-context-background);
+  }
+}
