+++ src/components/context/ContextBox.scss
@@ -1,116 +1,9 @@
-@import '../../design-patterns/colors/vars';
-@import '../../design-patterns/colors/colors-light';
 @import '../../design-patterns/spacing/vars';
-@import '../vars';
-
-$context-box-tip-size: $spacing-grid-size * 3;
-$context-box-tip-space: $spacing-grid-size * 2;
-
-.ax-context-box {
-  position: relative;
-  filter: drop-shadow($drop-shadow);
-}
-
-.ax-context-box__content {
-  border-radius: $component-border-radius;
-  background-color: var(--color-context-background);
-  color: var(--color-context-text);
-  overflow: hidden;
-}
-
-.ax-context-box__arrow {
-  position: absolute;
-  width: 0;
-  height: 0;
-  border-width: $context-box-tip-size;
-  border-style: solid;
-  border-color: transparent;
-}
-
-.ax-context-box--top {
-  margin-bottom: $context-box-tip-size;
-
-  .ax-context-box__content {
-    padding-top: $context-box-tip-space;
-    padding-bottom: $context-box-tip-space;
-  }
-}
-
-.ax-context-box--bottom {
-  margin-top: $context-box-tip-size;
-
-  .ax-context-box__content {
-    padding-top: $context-box-tip-space;
-    padding-bottom: $context-box-tip-space;
-  }
-}
-
-.ax-context-box--right {
-  margin-left: $context-box-tip-size;
-
-  .ax-context-box__content {
-    padding-left: $context-box-tip-space;
-  }
-}
-
-.ax-context-box--left {
-  margin-right: $context-box-tip-size;
-
-  .ax-context-box__content {
-    padding-right: $context-box-tip-space;
-  }
-}
 
 .ax-context-box--padding-small {
-  .ax-context-box__content {
-    padding: ($spacing-grid-size * 4);
-  }
+  padding: ($spacing-grid-size * 4);
 }
 
 .ax-context-box--padding-large {
-  .ax-context-box__content {
-    padding: ($spacing-grid-size * 6);
-  }
-}
-
-.ax-context-box--top,
-.ax-context-box--bottom {
-  .ax-context-box__arrow {
-    left: calc(50% - #{$context-box-tip-size});
-  }
-}
-
-.ax-context-box--right,
-.ax-context-box--left {
-  .ax-context-box__arrow {
-    top: calc(50% - #{$context-box-tip-size});
-  }
-}
-
-.ax-context-box--top {
-  .ax-context-box__arrow {
-    bottom: -($context-box-tip-size * 2);
-    border-top-color: var(--color-context-background);
-  }
-}
-
-.ax-context-box--right {
-  .ax-context-box__arrow {
-    left: -($context-box-tip-size * 2);
-    border-right-color: var(--color-context-background);
-  }
-}
-
-.ax-context-box--bottom {
-  .ax-context-box__arrow {
-    top: -($context-box-tip-size * 2);
-    border-bottom-color: var(--color-context-background);
-  }
-}
-
-.ax-context-box--left {
-  .ax-context-box__arrow {
-    right: -($context-box-tip-size * 2);
-    border-left-color: var(--color-context-background);
-  }
+  padding: ($spacing-grid-size * 6);
 }
