+++ src/components/context/ContextBox.test.js
@@ -17,30 +17,6 @@ describe('ContextBox', () => {
     expect(tree).toMatchSnapshot();
   });
 
-  it('renders top', () => {
-    const component = getComponent({ position: 'top' });
-    const tree = component.toJSON();
-    expect(tree).toMatchSnapshot();
-  });
-
-  it('renders right', () => {
-    const component = getComponent({ position: 'right' });
-    const tree = component.toJSON();
-    expect(tree).toMatchSnapshot();
-  });
-
-  it('renders bottom', () => {
-    const component = getComponent({ position: 'bottom' });
-    const tree = component.toJSON();
-    expect(tree).toMatchSnapshot();
-  });
-
-  it('renders left', () => {
-    const component = getComponent({ position: 'left' });
-    const tree = component.toJSON();
-    expect(tree).toMatchSnapshot();
-  });
-
   it('renders padding none', () => {
     const component = getComponent({ padding: 'none' });
     const tree = component.toJSON();
@@ -55,12 +31,6 @@ describe('ContextBox', () => {
 
   it('renders padding large', () => {
     const component = getComponent({ padding: 'large' });
-    const tree = component.toJSON();
-    expect(tree).toMatchSnapshot();
-  });
-
-  it('renders custom width', () => {
-    const component = getComponent({ width: '20rem' });
     const tree = component.toJSON();
     expect(tree).toMatchSnapshot();
   });
