+++ src/components/context/ContextMenu.scss
@@ -1,6 +1,29 @@
+@import '../../design-patterns/spacing/vars';
+
+$context-box-tip-space: $spacing-grid-size * 2;
+
 .ax-context-menu {
   margin: 0;
   padding: 0;
   list-style: none;
 }
 
+.ax-context--top,
+.ax-context--bottom {
+  .ax-context-menu {
+    padding-top: $context-box-tip-space;
+    padding-bottom: $context-box-tip-space;
+  }
+}
+
+.ax-context--right {
+  .ax-context-menu {
+    padding-left: $context-box-tip-space;
+  }
+}
+
+.ax-context--left {
+  .ax-context-menu {
+    padding-right: $context-box-tip-space;
+  }
+}
