+++ src/components/context/ContextMenuItem.scss
@@ -9,7 +9,7 @@
   color: var(--color-context-text);
 
   &:hover {
-    background-color: $color-ui-primary-1;
+    background-color: $color-ui-primary-2;
     color: $color-ui-light-2;
   }
 }
