+++ src/components/dropdown/example/button-group.js
@@ -4,6 +4,7 @@ import {
   Button,
   ButtonGroup,
   ButtonIcon,
+  Context,
   ContextMenu,
   ContextMenuItem,
   Dropdown,
@@ -27,11 +28,13 @@ export default class DropdownExample extends Component {
               </DropdownButton>
 
               <DropdownMenu>
-                <ContextMenu>
-                  <ContextMenuItem>Option 1</ContextMenuItem>
-                  <ContextMenuItem>Option 2</ContextMenuItem>
-                  <ContextMenuItem>Option 3</ContextMenuItem>
-                </ContextMenu>
+                <Context>
+                  <ContextMenu>
+                    <ContextMenuItem>Option 1</ContextMenuItem>
+                    <ContextMenuItem>Option 2</ContextMenuItem>
+                    <ContextMenuItem>Option 3</ContextMenuItem>
+                  </ContextMenu>
+                </Context>
               </DropdownMenu>
             </Dropdown>
           </ButtonGroup>
