+++ src/components/dropdown/example/icon-button.js
@@ -1,6 +1,7 @@
 import React, { Component } from 'react';
 import { Example, Snippet } from 'style-guide';
 import {
+  Context,
   ContextMenu,
   ContextMenuItem,
   Dropdown,
@@ -20,11 +21,13 @@ export default class DropdownExample extends Component {
             </DropdownButton>
 
             <DropdownMenu>
-              <ContextMenu>
-                <ContextMenuItem>Option 1</ContextMenuItem>
-                <ContextMenuItem>Option 2</ContextMenuItem>
-                <ContextMenuItem>Option 3</ContextMenuItem>
-              </ContextMenu>
+              <Context>
+                <ContextMenu>
+                  <ContextMenuItem>Option 1</ContextMenuItem>
+                  <ContextMenuItem>Option 2</ContextMenuItem>
+                  <ContextMenuItem>Option 3</ContextMenuItem>
+                </ContextMenu>
+              </Context>
             </DropdownMenu>
           </Dropdown>
         </Snippet>
